Update the whole config at once
Using one call to 'syncthing cli' for each config changes borks when updating
the GUI configuration (see [1]).

So instead I'm using jq to modify the config as desired and a single API call
to update the whole config at once. The GUI config changes still trigger a
restart of the internal http server, but by then the config change is done, and
usually we still manage to slip in the shutdown request before syncthing stops
accepting commands so it's super fast.

Unlike with the CLI, I think doing this way bypasses some of the validation,
but at this point I'm too jaded to check.

[1]: https://github.com/syncthing/syncthing/issues/8255