Remove syncthing's init container
It wasn't a real "init container", and caused issues when calling `docker
compose up` and the syncthing container was already up and running.

I'm not too happy about reimplementing/adapting upstream's Dockerfile to add
some customizations, and also I'm not too happy about having extra binaries in
the image that are only used for my custom initialization. But that's the price
to pay for having a syncthing container that "acts as if it was entirely
configurable through environment variables" (at restarts at least).

An alternative might be to rewrite the syncthing-start script in go, so I can
compile it in its own container, then get rid of all the extra dependencies for
the runtime container.