Upgrade syncthing to 1.23.7
The home directory can now be passed as environment variables, which simplifies
the wrapper script a tad.

Unfortunately it seems that the `syncthing cli` subcommand doesn't use it [1],
which makes it a bit annoying to run one-off commands in the container, but
doesn't affect the wrapper script (as we provide directly the gui address and
api key instead).

[1]: https://github.com/syncthing/syncthing/issues/9041