Upgrade authelia
The configuration was NOT migrated to the new authz endpoint and multi-domain
protection.

The reason is that it would require the introduction of at least a new
configuration file and enable the template filter to be able to keep the
configuration nice.

The legacy authentication endpoint still works (and is still required for
forced basic auth anyway), and the upgrade without migration will only produce
a single warning at startup for now.