Upgrade authelia to 4.39.1, migrate to new authz endpoints
Thanks to the config templating feature introduced in authelia 4.38,
environment variables can now be resolved directly by authelia, so I can ditch
the separate `.template` file that was used to anonymise the configuration in
VCS.

Sadly the new authz endpoints require the use of a configuration file, and it
is not possible to configure it using only the docker-compose file anymore.

The last notable change is to run the authelia container as a non-root user.