Upgrade syncthing to 2.0.9
The "default folder" concept was removed in syncthing 2.0.0.

While the DB migration is in progress, syncthing responds to all requests with
a fixed temporary message that breaks all API calls (incl. `syncthing cli`).
The migration can take longer than the timeout implemented for `syncthing cli`
commands, and if not handled properly at startup would make the startup script
exit with a failure, triggering a restart loop (as the migration always
restarts from scratch). The implemented workaround has the potential to get the
startup script stuck if `syncthing serve` has itself an issue (the initial
check doesn't have a timeout anymore).